requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.0.3
requests==2.31.0
numpy==1.24.3
streamlit==1.24.0
fastapi==0.100.0
uvicorn==0.23.2
beautifulsoup4==4.12.2